requirements.txt
streamlit
ultralytics
Pillow
opencv-python-headless
numpy